+++ lib/http.js
@@ -12,7 +12,6 @@ var Jsonic = require('jsonic')
 var Needle = require('needle')
 var Query = require('connect-query')
 var Timeout = require('connect-timeout')
-var TransportUtil = require('./transport-utils.js')
 
 // Declare internals
 var internals = {}
@@ -36,7 +35,7 @@ exports.listen = function (options, transportUtil) {
     seneca.log.debug('listen', listenOptions)
     var listener = app.listen(listenOptions.port, listenOptions.host)
 
-    TransportUtil.close(seneca, function (done) {
+    transportUtil.close(seneca, function (done) {
       listener.close()
       done()
     })
@@ -61,7 +60,7 @@ exports.client = function (options, transportUtil) {
           'seneca-id': data.id,
           'seneca-kind': 'req',
           'seneca-origin': seneca.id,
-          'seneca-track': TransportUtil.stringifyJSON(seneca, 'send-track', data.track || []),
+          'seneca-track': transportUtil.stringifyJSON(seneca, 'send-track', data.track || []),
           'seneca-time-client-sent': data.time.client_sent
         }
 
@@ -100,7 +99,7 @@ exports.client = function (options, transportUtil) {
         )
       })
 
-      TransportUtil.close(seneca, function (done) {
+      transportUtil.close(seneca, function (done) {
         done()
       })
     }
@@ -120,12 +119,12 @@ internals.setBody = function (seneca, transportUtil) {
       try {
         var bufstr = buf.join('')
 
-        var bodydata = bufstr.length ? TransportUtil.parseJSON(seneca, 'req-body', bufstr) : {}
+        var bodydata = bufstr.length ? transportUtil.parseJSON(seneca, 'req-body', bufstr) : {}
 
         if (Util.isError(bodydata)) {
-          var out = TransportUtil.prepareResponse(seneca, {})
+          var out = transportUtil.prepareResponse(seneca, {})
           out.input = bufstr
-          out.error = TransportUtil.error('invalid_json', { input: bufstr })
+          out.error = transportUtil.error('invalid_json', { input: bufstr })
           internals.sendResponse(seneca, transportUtil, res, out, {})
           return
         }
@@ -155,7 +154,7 @@ internals.trackHeaders = function (listenOptions, seneca, transportUtil) {
         id: req.headers['seneca-id'],
         kind: 'act',
         origin: req.headers['seneca-origin'],
-        track: TransportUtil.parseJSON(seneca, 'track-receive', req.headers['seneca-track']) || [],
+        track: transportUtil.parseJSON(seneca, 'track-receive', req.headers['seneca-track']) || [],
         time: {
           client_sent: req.headers['seneca-time-client-sent']
         },
@@ -189,10 +188,10 @@ internals.sendResponse = function (seneca, transportUtil, res, out, data) {
   var httpcode = 200
 
   if (out && out.res) {
-    outJson = TransportUtil.stringifyJSON(seneca, 'listen-web', out.res)
+    outJson = transportUtil.stringifyJSON(seneca, 'listen-web', out.res)
   } else if (out && out.error) {
     isError = true
-    outJson = TransportUtil.stringifyJSON(seneca, 'listen-web', out.error)
+    outJson = transportUtil.stringifyJSON(seneca, 'listen-web', out.error)
   }
 
   var headers = {
