+++ lib/tcp.js
@@ -3,16 +3,13 @@
 'use strict'
 
 // Load modules
-
 var Util = require('util')
 var Net = require('net')
 var Stream = require('stream')
 var Hoek = require('hoek')
 var Reconnect = require('reconnect-core')
-var TransportUtil = require('./transport-utils.js')
 
 // Declare internals
-
 var internals = {}
 
 exports.listen = function (options, transportUtil) {
@@ -26,9 +23,9 @@ exports.listen = function (options, transportUtil) {
       seneca.log.debug('listen', 'connection', listenOptions,
                        'remote', connection.remoteAddress, connection.remotePort)
       connection
-        .pipe(TransportUtil.jsonStreamParser(seneca))
+        .pipe(internals.jsonStreamParser(seneca, transportUtil))
         .pipe(internals.requestMessager(seneca, listenOptions, transportUtil))
-        .pipe(TransportUtil.jsonStringifyStream(seneca))
+        .pipe(internals.jsonStringifyStream(seneca, transportUtil))
         .pipe(connection)
 
       connection.on('error', function (err) {
@@ -53,7 +50,7 @@ exports.listen = function (options, transportUtil) {
 
     listener.listen(listenOptions.port, listenOptions.host)
 
-    TransportUtil.close(seneca, function (next) {
+    transportUtil.close(seneca, function (next) {
       listener.close()
       internals.closeConnections(connections, seneca)
       next()
@@ -77,9 +74,9 @@ exports.client = function (options, transportUtil) {
         connections.push(client)
 
         client
-          .pipe(TransportUtil.jsonStreamParser(seneca))
+          .pipe(internals.jsonStreamParser(seneca, transportUtil))
           .pipe(msger)
-          .pipe(TransportUtil.jsonStringifyStream(seneca))
+          .pipe(internals.jsonStringifyStream(seneca, transportUtil))
           .pipe(client)
       }).on('connect', function () {
           seneca.log.debug('client', type, 'connect', spec, topic, clientOptions)
@@ -98,7 +95,7 @@ exports.client = function (options, transportUtil) {
         msger.push(outmsg)
       })
 
-      TransportUtil.close(seneca, function (done) {
+      transportUtil.close(seneca, function (done) {
         clientconnect.disconnect()
 
         internals.closeConnections(connections, seneca)
@@ -117,9 +114,9 @@ internals.requestMessager = function (seneca, options, transportUtil) {
     var stream = this
 
     if (Util.isError(data)) {
-      var out = TransportUtil.prepareResponse(seneca, {})
+      var out = transportUtil.prepareResponse(seneca, {})
       out.input = data.input
-      out.error = TransportUtil.error('invalid_json', { input: data.input })
+      out.error = transportUtil.error('invalid_json', { input: data.input })
 
       stream.push(out)
       return next()
@@ -155,6 +152,59 @@ internals.closeConnections = function (connections, seneca) {
       seneca.log.error(e)
     }
   })
+}
+
+internals.jsonStreamParser = function (seneca, transportUtil) {
+  var parser = new Stream.Duplex({ objectMode: true })
+  parser.linebuf = []
+  parser._read = function () {}
+  parser._write = function (data, enc, callback) {
+    var str = '' + data
+    var endline = -1
+    var remain = 0
+
+    while ((endline = str.indexOf('\n', remain)) !== -1) {
+      this.linebuf.push(str.substring(remain, endline))
+      var jsonstr = this.linebuf.join('')
+
+      this.linebuf.length = 0
+      remain = endline + 1
+
+      if (jsonstr === '') {
+        return callback()
+      }
+
+      var outdata = transportUtil.parseJSON(seneca, 'stream', jsonstr)
+
+      if (outdata) {
+        this.push(outdata)
+      }
+    }
+
+    if (endline === -1) {
+      this.linebuf.push(str.substring(remain))
+    }
+
+    return callback()
+  }
+
+  return parser
+}
+
+internals.jsonStringifyStream = function (seneca, transportUtil) {
+  var stringify = new Stream.Duplex({ objectMode: true })
+  stringify._read = function () {}
+  stringify._write = function (data, enc, callback) {
+    var out = transportUtil.stringifyJSON(seneca, 'stream', data)
+
+    if (out) {
+      this.push(out + '\n')
+    }
+
+    callback()
+  }
+
+  return stringify
 }
 
 internals.reconnect = Reconnect(function () {
