+++ lib/transport-utils.js
@@ -3,9 +3,7 @@
 'use strict'
 
 // Load modules
-
 var Util = require('util')
-var Stream = require('stream')
 var _ = require('lodash')
 var Nid = require('nid')
 var Patrun = require('patrun')
@@ -14,7 +12,6 @@ var Jsonic = require('jsonic')
 var Eraro = require('eraro')
 
 // Declare internals
-
 var internals = {
   error: Eraro({
     package: 'seneca',
@@ -176,7 +173,7 @@ internals.Utils.prototype.handle_request = function (seneca, data, listen_option
     return respond({ input: data, error: internals.error('data_error') })
   }
 
-  var output = internals.Utils.prepareResponse(seneca, data)
+  var output = this.prepareResponse(seneca, data)
   var input = this.handle_entity(data.act)
 
   input.transport$ = {
@@ -476,7 +473,7 @@ internals.Utils.prototype.handle_entity = function (raw) {
   // legacy names
 internals.Utils.prototype.resolvetopic = internals.Utils.prototype.resolve_topic
 
-internals.Utils.close = function (seneca, closer) {
+internals.Utils.prototype.close = function (seneca, closer) {
   seneca.add('role:seneca,cmd:close', function (close_args, done) {
     var seneca = this
 
@@ -490,7 +487,7 @@ internals.Utils.close = function (seneca, closer) {
   })
 }
 
-internals.Utils.stringifyJSON = function (seneca, note, obj) {
+internals.Utils.prototype.stringifyJSON = function (seneca, note, obj) {
   if (!obj) {
     return
   }
@@ -502,7 +499,7 @@ internals.Utils.stringifyJSON = function (seneca, note, obj) {
   }
 }
 
-internals.Utils.parseJSON = function (seneca, note, str) {
+internals.Utils.prototype.parseJSON = function (seneca, note, str) {
   if (!str) {
     return
   }
@@ -516,7 +513,7 @@ internals.Utils.parseJSON = function (seneca, note, str) {
   }
 }
 
-internals.Utils.prepareResponse = function (seneca, input) {
+internals.Utils.prototype.prepareResponse = function (seneca, input) {
   return {
     id: input.id,
     kind: 'res',
@@ -530,63 +527,4 @@ internals.Utils.prepareResponse = function (seneca, input) {
   }
 }
 
-internals.Utils.jsonStreamParser = function (seneca) {
-  var parser = new Stream.Duplex({ objectMode: true })
-  parser.linebuf = []
-  parser._read = function () {}
-  parser._write = function (data, enc, callback) {
-    var str = '' + data
-    var endline = -1
-    var remain = 0
-
-    while ((endline = str.indexOf('\n', remain)) !== -1) {
-      this.linebuf.push(str.substring(remain, endline))
-      var jsonstr = this.linebuf.join('')
-
-      this.linebuf.length = 0
-      remain = endline + 1
-
-      if (jsonstr === '') {
-        return callback()
-      }
-
-      var outdata = internals.Utils.parseJSON(seneca, 'stream', jsonstr)
-
-      if (outdata) {
-        this.push(outdata)
-      }
-    }
-
-    if (endline === -1) {
-      this.linebuf.push(str.substring(remain))
-    }
-
-    return callback()
-  }
-
-  return parser
-}
-
-internals.Utils.jsonStringifyStream = function (seneca) {
-  var stringify = new Stream.Duplex({ objectMode: true })
-  stringify._read = function () {}
-  stringify._write = function (data, enc, callback) {
-    var out = internals.Utils.stringifyJSON(seneca, 'stream', data)
-
-    if (out) {
-      this.push(out + '\n')
-    }
-
-    callback()
-  }
-
-  return stringify
-}
-
-internals.Utils.error = internals.error
-
-// Needed for transport exports of utils
-internals.Utils.prototype.close = internals.Utils.close
-internals.Utils.prototype.stringifyJSON = internals.Utils.stringifyJSON
-internals.Utils.prototype.parseJSON = internals.Utils.parseJSON
-internals.Utils.prototype.prepareResponse = internals.Utils.prepareResponse
+internals.Utils.prototype.error = internals.error
